tests/wpt/web-platform-tests/css/css-forced-color-adjust/inheritance.html

<!DOCTYPE html>
<title>Inheritance of CSS Forced Color Adjust properties</title>
<link rel="help" href="https://www.w3.org/TR/css-color-adjust-1/#forced-color-adjust-prop">
<meta name="assert" content="Properties inherit according to the spec.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
<div id="container">
  <div id="target"></div>
</div>
<script>
  assert_inherited('forced-color-adjust', 'auto', 'none');
</script>
